requirements.txt
# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# AI/LLM Libraries
google-generativeai>=0.8.0
openai>=1.12.0
anthropic>=0.18.0
langchain>=0.1.0
langchain-google-genai>=1.0.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0

# A2A Protocol
a2a-sdk>=0.1.0  # Google A2A SDK (when available)

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0

# HTTP & Async
httpx>=0.26.0
aiohttp>=3.9.0

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.6
pyyaml>=6.0.0
tenacity>=8.2.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
httpx>=0.26.0

# Development
black>=24.0.0
isort>=5.13.0
mypy>=1.8.0
ruff>=0.2.0
